Refatorar sessao ingresso atracao

+++ src/main/java/br/ufg/inf/pitanga/interfaces/dao/InterfaceIngressoDao.java
@@ -17,7 +17,7 @@ public interface InterfaceIngressoDao {
 
     public void alterarIngresso(Ingresso umIngresso);
 
-    public Ingresso buscarPorId(int Id);
+    public Ingresso buscarPorId(Long Id);
 
     public List<Ingresso> buscaPorCliente(Cliente umCliente);
